Fix : [코드 리뷰 반영] Comment Entity - root entity 공통 코드 삭제 (생성시각, 수정시각) - Repository 타입 BigInteger -> Long으로 변경 - 클래스 멤버변수명 Snake -> Camel 스타일로 변경 - 컬럼 length 지정

+++ src/main/java/notai/comment/domain/CommentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, BigInteger> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 
 }
